requirements.txt
accelerate==1.3.0
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bibtexparser==1.4.3
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
docarray==0.40.0
docutils==0.21.2
durationpy==0.9
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.2
executing==2.1.0
faiss-cpu==1.10.0
fake-useragent==2.0.3
fastapi==0.115.8
fastjsonschema==2.21.1
feedparser==6.0.11
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.17.0
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==25.1.24
fonttools==4.56.0
fqdn==1.5.1
free_proxy==1.1.3
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==3.10.0
google_search_results==2.4.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
kubernetes==32.0.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-groq==0.2.4
langchain-huggingface==0.1.2
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.6
libmagic==1.0
looseversion==1.3.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mutagen==1.47.0
mypy-extensions==1.0.0
narwhals==1.25.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
neo4j==5.28.1
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.7
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.20.1
openai==1.61.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20240706
# pdfplumber==0.11.5
pi_heif==0.21.0
pikepdf==9.5.2
pillow==11.1.0
platformdirs==4.3.6
plum-dispatch==1.7.4
plumbum==1.9.0
ply==3.11
poppler-utils==0.1.0
posthog==3.11.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
prov==2.0.1
psutil==6.1.1
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.3
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.3.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.8
python-json-logger==3.2.1
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.1
pywin32==308
pywinpty==2.0.14
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.12.1
rdflib==6.3.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
roman-numerals-py==3.1.0
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scholarly==1.7.11
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.28.1
Send2Trash==1.8.3
sentence-transformers==3.4.1
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.2.1
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
Spire.Pdf==10.12.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.45.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
terminado==0.18.1
tesseract==0.1.3
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.14
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.48.2
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.23
unstructured-client==0.30.1
unstructured-inference==0.8.7
unstructured.pytesseract==0.3.13
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
watermark==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.18.3
yt-dlp==2025.1.26
zipp==3.21.0
zstandard==0.23.0
